Merges with SCIDI to help integrating build process with it Remove some gcc specific flags from Makefile.am Rename some header defines size_t vs. int fixups opensc.h: Define inline as null for other compilers than gcc, for now Port pam_pkcs15 to compile for Solaris and HP-UX, untested Fix compiler warnings
OpenSC now compiles cleanly for Tru64, AIX and HP-UX.

The only problem is the tools using getopt_long() (GNU extension),
to be fixed later..


git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@120 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/libopensc/card-setcos.c
@@ -44,7 +44,7 @@ static int setec_match_card(struct sc_card *card)
 
 	for (i = 0; setec_atrs[i] != NULL; i++) {
 		u8 defatr[SC_MAX_ATR_SIZE];
-		int len = sizeof(defatr);
+		size_t len = sizeof(defatr);
 		const char *atrp = setec_atrs[i];
 
 		if (sc_hex_to_bin(atrp, defatr, &len))
